requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.9.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
nvidia-cuda-runtime-cu12==12.6.77
openai==1.51.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sympy==1.13.3
tensorrt==10.5.0
tensorrt-cu12==10.5.0
tensorrt-cu12_bindings==10.5.0
tensorrt-cu12_libs==10.5.0
tensorrt-dispatch==10.5.0
tensorrt-dispatch-cu12==10.5.0
tensorrt-lean==10.5.0
tensorrt-lean-cu12==10.5.0
tensorrt_dispatch-cu12_bindings==10.5.0
tensorrt_dispatch-cu12_libs==10.5.0
tensorrt_lean-cu12_bindings==10.5.0
tensorrt_lean-cu12_libs==10.5.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.13
ultralytics-thop==2.0.9
urllib3==2.2.3
Werkzeug==3.0.4
